+++ lib/flutter_easy_button.dart
@@ -1,1 +1,1 @@
-export 'src/custom_filled_button.dart';
+export 'src/custom_button.dart';